+++ main.go
@@ -2,14 +2,12 @@ package main
 
 // Import our dependencies. We'll use the standard HTTP library as well as the gorilla router for this app
 import (
-	"encoding/json"
-	"fmt"
 	"net/http"
 	"os"
 
-	"github.com/darenliang/jikan-go"
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
+	"github.comoraora-io/oraora-be/seasons.go"
 )
 
 // Name of the anime
@@ -24,7 +22,7 @@ func main() {
 	// On the default page we will simply serve our static index page.
 	r.Handle("/", http.FileServer(http.Dir("./views/")))
 	r.Handle("/status", StatusHandler).Methods("GET")
-	r.Handle("/anime", AnimeHandler).Methods("GET")
+	r.Handle("/anime", seasons.AnimeHandler).Methods("GET")
 
 	// We will setup our server so we can serve static assest like images, css from the /static/{file} route
 	r.PathPrefix("/static/").Handler(http.StripPrefix("/static/", http.FileServer(http.Dir("./static/"))))
@@ -37,60 +35,6 @@ func main() {
 // StatusHandler ...
 var StatusHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte("API is up and running"))
-})
-
-// AnimeHandler reports the data in raw JSON
-var AnimeHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
-	// Gets all the anime information data based on year and season
-	season, _ := jikan.GetSeason(jikan.Season{
-		Year:   2019,
-		Season: "fall",
-	})
-
-	// Starting with an empty slice of strings
-	var animeSlice []string
-
-	//TODO: figure out a more elegant way of getting number of anime in a given season
-	//Can't natively index type interface{}, so convert to []interface{}, and then index, with output of type map[string]interface{}
-	for i := 0; i < 182; i++ {
-
-		anime := season["anime"].([]interface{})[i].(map[string]interface{})
-
-		//Marshal the data into json
-
-		animeData, err := json.Marshal(anime)
-
-		if err != nil {
-			fmt.Println(err.Error())
-			return
-		}
-
-		// type Name declared as a struct at the top
-		var title Name
-
-		// Unmarshal the JSON to only grab the title of the anime
-		json.Unmarshal([]byte(string(animeData)), &title)
-		if err != nil {
-			fmt.Println(err.Error())
-			return
-		}
-
-		// Append title of the anime to empty slice of strings animeSlice
-		animeSlice = append(animeSlice, title.Title)
-
-	}
-
-	// Marshal the entire slice of strings back into JSON
-	json.Marshal(animeSlice)
-	titleJSON, err := json.Marshal(animeSlice)
-	if err != nil {
-		fmt.Println(err.Error())
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.Write([]byte(titleJSON))
 })
 
 // NotImplemented ...
